init changes
